Added RequestReceiver class as a mixin for endpoint and service. Get returns with 'values' field without a specifier.
